make utils for geo and datetime
