+++ featuretools/primitives/transform/datetime/date_to_holiday.py
@@ -3,7 +3,7 @@ from woodwork.column_schema import ColumnSchema
 from woodwork.logical_types import Categorical, Datetime
 
 from featuretools.primitives.core.transform_primitive import TransformPrimitive
-from featuretools.primitives.utils import HolidayUtil
+from featuretools.primitives.transform.datetime.utils import HolidayUtil
 
 
 class DateToHoliday(TransformPrimitive):
