+++ featuretools/primitives/transform/geo/cityblock_distance.py
@@ -4,7 +4,7 @@ from woodwork.column_schema import ColumnSchema
 from woodwork.logical_types import Double, LatLong
 
 from featuretools.primitives.core.transform_primitive import TransformPrimitive
-from featuretools.primitives.utils import _haversine_calculate
+from featuretools.primitives.transform.geo.utils import _haversine_calculate
 
 
 class CityblockDistance(TransformPrimitive):
