+++ featuretools/primitives/transform/geo/haversine.py
@@ -3,7 +3,7 @@ from woodwork.column_schema import ColumnSchema
 from woodwork.logical_types import LatLong
 
 from featuretools.primitives.core.transform_primitive import TransformPrimitive
-from featuretools.primitives.utils import _haversine_calculate
+from featuretools.primitives.transform.geo.utils import _haversine_calculate
 
 
 class Haversine(TransformPrimitive):
